### 2017-01-13 #162 @whedon
[PRE REVIEW]: flusight
**Submitting author:** @lepisma (<a href="http://orcid.org/0000-0001-7137-0728">Abhinav Tushar</a>)
**Repository:** <a href="https://github.com/reichlab/flusight" target ="_blank">https://github.com/reichlab/flusight</a>
**Version:** v1.1.0
