+++ src/main.js
@@ -5,7 +5,6 @@ import App from './App'
 import store from './store'
 
 import 'bulma/css/bulma.css'
-import 'font-awesome/css/font-awesome.css'
 import './assets/fakeLoader.css'
 import './assets/favicon.ico'
 import './assets/analytics.js'
