### 2017-03-06 fd66524
Revert "Merge branch 'master' of github.com:reichlab/flusight"
This reverts commit 7a90c65a1abe4b83070a67508ae7b09a1491af87, reversing
changes made to 6742662be48d5b57e529eb10042c053d93c18d11.
