+++ src/modules/timechart.js
@@ -1,533 +1,0 @@
-// Chart plotting functions
-
-import * as util from './utils/timechart'
-import * as marker from './markers/timechart'
-import textures from 'textures'
-import * as mmwr from 'mmwr-week'
-import * as d3 from 'd3'
-
-export default class TimeChart {
-  constructor (elementId, weekHook) {
-    // Get div dimensions
-    let footBB = d3.select('.footer').node().getBoundingClientRect()
-    let chartBB = d3.select('#' + elementId).node().getBoundingClientRect()
-
-    let divWidth = chartBB.width
-    let divHeight = window.innerHeight - chartBB.top - footBB.height
-
-    // Padding offsets
-    divHeight -= 20
-
-    // Limits
-    divHeight = Math.min(Math.max(350, divHeight), 550)
-
-    // Height of onset panel above x axis
-    let onsetHeight = 30
-
-    // Create blank chart
-    let margin = {
-      top: 5, right: 50, bottom: 70 + onsetHeight, left: 40
-    }
-    let width = divWidth - margin.left - margin.right
-    let height = divHeight - margin.top - margin.bottom
-
-    // Initialize scales and stuff
-    let xScale = d3.scaleLinear()
-        .range([0, width])
-    let yScale = d3.scaleLinear()
-        .range([height, 0])
-    let xScaleDate = d3.scaleTime()
-        .range([0, width])
-
-    // Add svg
-    let svg = d3.select('#' + elementId).append('svg')
-        .attr('width', width + margin.left + margin.right)
-        .attr('height', height + margin.top + margin.bottom)
-        .append('g')
-        .attr('transform', 'translate(' + margin.left + ',' + margin.top + ')')
-
-    // Add tooltips
-    this.chartTooltip = d3.select('#chart-tooltip')
-      .style('display', 'none')
-
-    this.legendTooltip = d3.select('#legend-tooltip')
-      .style('display', 'none')
-
-    // Save variables
-    this.svg = svg
-    this.xScale = xScale
-    this.yScale = yScale
-    this.xScaleDate = xScaleDate
-    this.height = height
-    this.width = width
-    this.onsetHeight = onsetHeight
-    this.weekHook = weekHook
-
-    // Add axes
-    this.setupAxes()
-
-    // Add marker primitives
-    this.timerect = new marker.TimeRect(this)
-
-    this.onsetTexture = textures.lines()
-      .lighter()
-      .strokeWidth(0.5)
-      .size(8)
-      .stroke('#ccc')
-    svg.call(this.onsetTexture)
-
-    // Paint the onset panel
-    this.paintOnsetOffset()
-
-    // Add overlays and other mouse interaction items
-    this.setupOverlay()
-
-    // Axis at top of onset panel
-    this.setupReverseAxis()
-
-    this.history = new marker.HistoricalLines(this)
-    this.baseline = new marker.Baseline(this)
-    this.actual = new marker.Actual(this)
-    this.observed = new marker.Observed(this)
-    this.predictions = []
-
-    // Hard coding confidence values as of now
-    // This and currently selected id should ideally go in the vuex store
-    this.confidenceIntervals = ['90%', '50%']
-    this.cid = 1 // Use 50% as default
-
-    // Legend toggle state
-    this.historyShow = true
-    this.predictionsShow = {}
-  }
-
-  /**
-   * Setup axes
-   */
-  setupAxes () {
-    let svg = this.svg
-    let width = this.width
-    let height = this.height
-    let onsetHeight = this.onsetHeight
-
-    // Keep onset panel between xaxis and plot
-    let xAxisPos = height + onsetHeight
-
-    let infoTooltip = d3.select('#info-tooltip')
-        .style('display', 'none')
-
[…]
