### 2017-03-12 a70ac1a
Fix no-data issue when lag is above limit for a week range
