### 2017-03-12 a4bbdbf
Add progressbar in scripts
