### 2017-03-26 8e65218
Remove unused getters, actions
