+++ src/store/mutation-types.js
@@ -10,7 +10,6 @@ export const UPDATE_SELECTED_WEEK = 'UPDATE_SELECTED_WEEK'
 // switch module
 export const UPDATE_SELECTED_REGION = 'UPDATE_SELECTED_REGION'
 export const UPDATE_SELECTED_SEASON = 'UPDATE_SELECTED_SEASON'
-export const TOGGLE_PANEL = 'TOGGLE_PANEL'
 export const TOGGLE_CHOROPLETH_RELATIVE = 'TOGGLE_CHOROPLETH_RELATIVE'
 
 // intro module
@@ -20,6 +19,3 @@ export const RESET_INTRO_POINTER = 'RESET_INTRO_POINTER'
 export const HIDE_INTRO = 'HIDE_INTRO'
 export const SHOW_INTRO = 'SHOW_INTRO'
 export const APPEND_INTRO_ITEM = 'APPEND_INTRO_ITEM'
-
-// models module
-export const UPDATE_SELECTED_STAT = 'UPDATE_SELECTED_STAT'
