### 2017-03-26 256b171
Move intro initialization code to intro.vue
