### 2017-03-26 ee3f8d1
Remove unnecessary key from timeChartData
