### 2017-03-26 d47e994
Fix linting error
