### 2017-03-30 0873dc2
Getter for timePoints
