### 2017-03-30 0398aaa
Return timepoints separately
