### 2017-03-31 9a418f9
Passing CIs and statsMeta in getters
