### 2017-04-05 #162 @arfon
@rneher @robohyun66 @sbfnk would one of you be willing to review this submission for JOSS? You can read more about what this entails here: http://joss.theoj.org/about#reviewer_guidelines
