### 2017-04-06 #162 @rneher
I'll take it on. seems like an interesting idea/project.
