### 2017-04-06 #231 @whedon
[REVIEW]: flusight
**Submitting author:** @lepisma (<a href="http://orcid.org/0000-0001-7137-0728">Abhinav Tushar</a>)
**Repository:** <a href="https://github.com/reichlab/flusight" target ="_blank">https://github.com/reichlab/flusight</a>
**Version:** v1.3.3

- [x] **Authorship:** Has the submitting author (lepisma) made major contributions to the software?
