### 2017-04-06 #162 @arfon
@rneher please head over to #231 for the actual review.
