### 2017-04-06 #231 @arfon
:wave: @rneher - please take a look at the reviewer guidelines (http://joss.theoj.org/about#reviewer_guidelines) and update the checklist above as you go. If you have any questions for me or the submitting author just ask us here.
