### 2017-04-10 a8ed503
New getters for distribution charts
