### 2017-04-10 d6abb46
Convert stats to list in data.json build process
