### 2017-04-10 a48be0d
return series instead of oneWk, twoWk ...
