### 2017-04-11 4854863
Fix issue with -1s in prediction
