### 2017-04-11 d748b5b
Convert week point predictions to indices in parse step
