+++ src/store/mutation-types.js
@@ -12,6 +12,8 @@ export const UPDATE_SELECTED_WEEK = 'UPDATE_SELECTED_WEEK'
 export const UPDATE_SELECTED_REGION = 'UPDATE_SELECTED_REGION'
 export const UPDATE_SELECTED_SEASON = 'UPDATE_SELECTED_SEASON'
 export const TOGGLE_CHOROPLETH_RELATIVE = 'TOGGLE_CHOROPLETH_RELATIVE'
+export const DISPLAY_TIMECHART = 'DISPLAY_TIMECHART'
+export const DISPLAY_DISTRIBUTIONCHART = 'DISPLAY_DISTRIBUTIONCHART'
 
 // intro module
 export const INCREMENT_INTRO_POINTER = 'INCREMENT_INTRO_POINTER'
