### 2017-04-11 cc49e83
Array compression/decompression functions
