### 2017-04-11 4f69d9d
Update decompressed array values iteratively
