### 2017-04-12 2e205af
Use newer foresight. Temporarily disable schema test
