+++ src/store/data.js
@@ -1,0 +1,1 @@
+export { data, updateTime } from '!json!../assets/data.json'
