+++ src/store/mutation-types.js
@@ -1,6 +1,7 @@
 // All the possible mutations for the application
 
 // Init mutations
+export const SET_DATA = 'SET_DATA'
 export const SET_TIMECHART = 'SET_TIMECHART'
 export const SET_CHOROPLETH = 'SET_CHOROPLETH'
 export const SET_DISTRIBUTIONCHART = 'SET_DISTRIBUTIONCHART'
