### 2017-04-24 #231 @rneher
flusight is a useful and intuitive visualization of flu prediction data. I managed to run a local dev server without problems. I don't see anything fundamental preventing acceptance. A few minor things

  - references don't render in the `paper.md`
  - a few additional sentences on how these predictions are visualized and evaluated now would help. Are there any similar efforts?
  - the tool tip for the prediction should explicitly mention "1 week ahead", "2 week ahead" etc.
  - historical traces are almost invisible, consider making them a little darker
  - an `about` page would help. Explain what is being predicted, what is used as input etc
  - the input prediction files have prediction for HHS regions but the website only offers US national. Is this intended?
