### 2017-04-24 #231 @arfon
Thanks @rneher. @lepisma - please let me know when you've had a chance to review this feedback.

On this point:

> references don't render in the paper.md

These will be compiled in at then end when we produce the PDF with Pandoc.
