### 2017-04-26 b963766
Allow number of points to keep while compressing array
