### 2017-04-26 8774d74
Better loader
