### 2017-04-26 eea9c26
Change getters to work with no data loaded
