### 2017-04-26 3a1fa63
Use 12 points in compression
