+++ scripts/utils.js
@@ -159,7 +159,8 @@ const regionFilter = (data, region) => {
 
     let compressedBins = null
     if (d.bins) {
-      // compressedBins = compressArray(d.bins.map(b => b[2]), 0.1)
+      // Using 12 points gives cushion for 33 bins targets
+      compressedBins = compressArray(d.bins.map(b => b[2]), 12)
     }
     if (wAIdx > -1) {
       filtered.series[wAIdx] = {
