### 2017-04-26 7499f3a
Move initial pointer finding code to flusight
