### 2017-04-26 ee5531c
Fix issue with pointer update
