### 2017-04-26 0caab75
Remove stray key
