### 2017-04-27 45a3a7f
Move chart init inside actions
