### 2017-04-27 c5994c7
Show both right charts on same div
