### 2017-04-27 d8751cc
On the fly curve generation
