### 2017-04-28 6e79fee
Fix decompression output to reflect real range
