### 2017-04-28 cf3289a
Remove decompression function
