+++ scripts/utils.js
@@ -37,43 +37,6 @@ const compressArray = (array, cRatio) => {
 }
 
 /**
- * Regenerate array in the compressed representation
- * Assume its a probability distribution and so it sums to one
- */
-const deCompressArray = compArray => {
-  let array = Array(Math.max(...compArray.map(i => i[0]))).fill(0)
-
-  let clamped = []
-  // Fill in given values
-  compArray.forEach(item => {
-    array[item[0]] = item[1]
-    clamped.push(item[0])
-  })
-
-  let epsilon = 0.00000001 // Try to acheive 1e-8 error
-  let alpha = 1
-  let maxIter = 100
-  let sum, error
-
-  for (let i = 0; i < maxIter; i++) {
-    sum = array.reduce((a, b) => (a + b), 0)
-    error = 1 - sum
-    if (Math.abs(error) < epsilon) {
-      break
-    } else {
-      // Neighbour filling
-      for (let j = 1; j < array.length - 1; j++) {
-        if (clamped.indexOf(j) === -1) {
-          array[j] += alpha * error * (array[j - 1] + array[j + 1]) / 2
-        }
-      }
-    }
-  }
-
-  return array
-}
-
-/**
  * Return list of weekStamps in given mmwr season
  */
 const seasonToWeekStamps = season => {
