### 2017-04-28 8015b33
Increase points for compression
