### 2017-04-28 c1b9526
Limit movements while in distributionChart mode
