### 2017-04-28 e17cc8d
Reset week index to valid value after redraws
