### 2017-04-30 df66343
Use grouped bins for distribution
