### 2017-04-30 05f6edc
Fix issue with grouping
