### 2017-05-01 78c05aa
Comment out array decompression
