### 2017-05-01 4e8a665
Set hooks on distributionChart
