### 2017-05-01 #231 @lepisma
Thanks for your comments @rneher !

For points
> - a few additional sentences on how these predictions are visualized and evaluated now would help. Are there any similar efforts?
> - an about page would help. Explain what is being predicted, what is used as input etc

Will add a github wiki page about this and link to the application.

> - the tool tip for the prediction should explicitly mention "1 week ahead", "2 week ahead" etc.
> - historical traces are almost invisible, consider making them a little darker

Working on these now.

> - the input prediction files have prediction for HHS regions but the website only offers US national. Is this intended?

The region to view predictions for can be selected by clicking on the map. This is mentioned in the interactive help (when we press the <kbd>help</kbd> button on the top right of site). Should this be more clearly mentioned somewhere else (like a dropdown menu)?
