### 2017-05-01 869d7d6
Add explaination about axis type in distribution panels
