+++ config.yaml
@@ -4,6 +4,7 @@ branding:
   parent: "reichlab"
   description: "Real-time influenza forecasts from Reich Lab @ UMass Amherst"
   imageUrl: "https://cdn.rawgit.com/reichlab/flusight/master/screen.png"
+  aboutUrl: "https://github.com/reichlab/flusight/wiki/About"
   appUrl: "https://reichlab.github.io/flusight"
   sourceUrl: "https://github.com/reichlab/flusight"
   parentUrl: "https://reichlab.github.io"
