+++ README.md
@@ -11,7 +11,7 @@ repository (which uses CDC's prediction format) or by the user.
 
 The live version associated with this github
 repository
-at [https://reichlab.github.io/flusight/](http://reichlab.io/flusight/)
+at [http://reichlab.io/flusight/](http://reichlab.io/flusight/)
 displays predictions for CDC FluSight Challenge
 by [Reich Lab](http://reichlab.io/) at UMass Amherst. To know more about it, go to the [about](https://github.com/reichlab/flusight/wiki/About) page in the [wiki](https://github.com/reichlab/flusight/wiki).
 
