### 2017-05-01 #231 @lepisma
Updates:

- There is a simple about page [here](https://github.com/reichlab/flusight/wiki/About) and is linked in the application (navbar).
- Chart tooltips show `x wks ahead` text.
- Historical traces are darker.
- The region to view predictions for can be selected via a dropdown (as well as the original method of clicking on the map).
