### 2017-05-05 #231 @nickreich
In response to the comment

> a few additional sentences on how these predictions are visualized and evaluated now would help. Are there any similar efforts?

I just committed a few additional links in the first paragraph of the About page.

We have completed our round of revisions, based on comments from @rneher and will wait to hear from JOSS about next steps. Thanks for the review!
