### 2017-05-10 #231 @lepisma
We have updated the version to [v1.3.3](https://github.com/reichlab/flusight/releases/tag/v1.3.3).
DOI : [https://doi.org/10.5281/zenodo.573580](https://doi.org/10.5281/zenodo.573580).
