### 2017-05-10 #231 @arfon
@rneher - many thanks for your review here ✨

@lepisma - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00231 ⚡️ 🚀 :boom:
